Start removing circular deps
