+++ apps/web/package.json
@@ -8,7 +8,8 @@
     "dev:test": "alinea dev -- next dev",
     "build": "yarn build:alinea && alinea build -- next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "seed": "alinea build -- tsx src/seed.ts"
   },
   "dependencies": {
     "@monaco-editor/react": "^4.4.5",
@@ -32,6 +33,7 @@
     "eslint": "8.39.0",
     "eslint-config-next": "13.4.0",
     "raw-loader": "^4.0.2",
+    "tsx": "^4.7.0",
     "typescript": "^5.3.3"
   },
   "exports": {
