+++ src/core/driver/DefaultDriver.server.tsx
@@ -2,7 +2,6 @@ import {Database} from 'alinea/backend'
 import {createStore} from 'alinea/backend/Store'
 import {EntryResolver} from 'alinea/backend/resolver/EntryResolver'
 import {createCloudHandler} from 'alinea/cloud/server/CloudHandler'
-import {base64} from 'alinea/core/util/Encoding'
 import PLazy from 'p-lazy'
 import {CMS} from '../CMS.js'
 import {Client} from '../Client.js'
@@ -10,6 +9,7 @@ import {Config} from '../Config.js'
 import {Connection} from '../Connection.js'
 import {Resolver} from '../Resolver.js'
 import {Realm} from '../pages/Realm.js'
+import {base64} from '../util/Encoding.js'
 import {Logger} from '../util/Logger.js'
 
 const store = PLazy.from(async () => {
