+++ src/core/media/CreatePreview.browser.ts
@@ -1,7 +1,7 @@
-import {base64} from 'alinea/core/util/Encoding'
 import {rgba, toHex} from 'color2k'
 import smartcrop from 'smartcrop'
 import {rgbaToThumbHash, thumbHashToAverageRGBA} from 'thumbhash'
+import {base64} from '../util/Encoding.js'
 import type {ImagePreviewDetails} from './CreatePreview.js'
 
 export {ImagePreviewDetails}
