+++ src/core/media/MediaSchema.browser.tsx
@@ -1,4 +1,3 @@
-import {type} from 'alinea/core'
 import {IcRoundPermMedia} from 'alinea/ui/icons/IcRoundPermMedia'
 import {MediaExplorer} from '../../dashboard/view/MediaExplorer.js'
 import {FileEntry} from '../../dashboard/view/media/FileEntry.js'
@@ -6,31 +5,31 @@ import {
   FileSummaryRow,
   FileSummaryThumb
 } from '../../dashboard/view/media/FileSummary.js'
-import {Meta} from '../Meta.js'
+import {type} from '../Type.js'
 import * as config from './MediaSchema.js'
 
 export type MediaLibrary = config.MediaLibrary
 export const MediaLibrary: typeof config.MediaLibrary = type(
   'Media directory',
   {
-    ...config.MediaLibrary,
-    [Meta]: {
-      isContainer: true,
-      contains: ['MediaLibrary'],
-      view: MediaExplorer,
-      icon: IcRoundPermMedia
+    isContainer: true,
+    contains: ['MediaLibrary'],
+    view: MediaExplorer,
+    icon: IcRoundPermMedia,
+    fields: {
+      ...config.MediaLibrary
     }
   }
 )
 
 export type MediaFile = config.MediaFile
 export const MediaFile: typeof config.MediaFile = type('File', {
-  ...config.MediaFile,
-  [Meta]: {
-    isHidden: true,
-    summaryRow: FileSummaryRow,
-    summaryThumb: FileSummaryThumb,
-    view: FileEntry
+  isHidden: true,
+  summaryRow: FileSummaryRow,
+  summaryThumb: FileSummaryThumb,
+  view: FileEntry,
+  fields: {
+    ...config.MediaFile
   }
 })
