--- src/core/pages/Condition.ts
@@ -1,3 +1,0 @@
-import {Expr, HasExpr} from './Expr.js'
-
-export type Condition = Expr<boolean> | HasExpr<boolean>
