+++ src/core/pages/Cursor.ts
@@ -1,9 +1,9 @@
 import {array, boolean, enums, number, object, string} from 'cito'
 import {Type} from '../Type.js'
 import {entries} from '../util/Objects.js'
-import {Condition} from './Condition.js'
 import {createExprData} from './CreateExprData.js'
 import {createSelection} from './CreateSelection.js'
+import type {Condition} from './Expr.js'
 import {EV, Expr, HasExpr} from './Expr.js'
 import {BinaryOp, ExprData} from './ExprData.js'
 import {Projection} from './Projection.js'
