+++ src/core/pages/TargetData.ts
@@ -1,5 +1,5 @@
-import {TypeTarget} from 'alinea/core/Type'
 import {object, string, type} from 'cito'
+import {TypeTarget} from '../Type.js'
 
 const TT = type(
   (value): value is TypeTarget => value && typeof value === 'object'
