+++ src/picker/entry/EntryPicker.ts
@@ -5,7 +5,7 @@ import {Label} from 'alinea/core/Label'
 import {Reference} from 'alinea/core/Reference'
 import {Type, type} from 'alinea/core/Type'
 import {MediaFile} from 'alinea/core/media/MediaSchema'
-import {Condition} from 'alinea/core/pages/Condition'
+import type {Condition} from 'alinea/core/pages/Expr'
 import {Projection} from 'alinea/core/pages/Projection'
 import {RecordShape} from 'alinea/core/shape/RecordShape'
 import {ScalarShape} from 'alinea/core/shape/ScalarShape'
